Created analytical software (read description for more)
Updated README.md.
Updated the database model to reflect, information needed on the user for calculations.
Updated the data collection software to include the time (duration) and changed it so the tester can put in their name so the file names have meaning.
Created rowinglib.py. there isn't much their yet, but it will be a module of helper functions for the main program.
Created the data analysis software. the Data class takes the 2 csv files and some other stored data on the user. it analyses the data and returns the formatted html for the graphs to embed on the users dashboard