created software to find a baseline for the EMG
updated the data collection software to retrieve the raw data rather than normalized.
updated the EMG class in devices.py to properly read the pin as analog using a library that works with the MCP3008 ADC.
created find_baseline.py to inerpret the emg data collected from subjects to find a baseline to normalize the readings. it uses a high-pass and low-pass filter on the data. fixes the offset, and rectify's the data. it returns one value as a baseline.
updated README.md to reflect new information on METs calculation
updated data analysis to calulate MET's and return intensity plot. also fixed list comprehension order